requirements.txt
PyInputPlus
requests
xlwt
numpy
linetool
matplotlib
Pillow
yfinance
tkintermapview
psycopg2-binary
pyinstaller